Global logger that wraps a consumer for common logging
The default global logger is effectively just System.out/err wrapped.
This logger uses the same implementation that was made in
MinecraftMaven, with indentations and some standard levels. The wrap is
in case other projects want to use something other than sysout, like a
Gradle project or task's own logger.

Also, Logging Utils now targets Java 8 so that it can be used in Java
Version if necessary. There's not a lot of code in this though so who
cares.